+++ package.json
@@ -14,6 +14,7 @@
     "react-dom": "^16.13.1",
     "react-full-screen": "^0.3.1",
     "react-magnifier": "^3.0.4",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "^3.4.3",
     "react-slick": "^0.27.11",
     "slick-carousel": "^1.8.1"
